move minimap, and refactor states to inner components
